+++ CHANGELOG.md
@@ -16,6 +16,44 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - N/A
 
+## [1.6.2]
+
+### Added
+- Template for orders in `src/templates`
+
+### Changed
+- N/A
+
+### Fixed
+- enpoint logic for orders enpoint in `src/app/v1/apps/orders/router.py`
+
+## [1.6.1]
+
+### Added 2026-05-24
+- Added placeholder orders template in `src/templates/orders.html`
+- `router.py` in `src/app/v1/orders/`
+- orders endpoint in `src/app/v1/orders/router.py`
+
+### Changed
+- N/A
+
+### Fixed
+- N/A
+
+## [1.6.0]
+
+### Added
+- auth flow in `src/app/v2/auth/auth.py`
+- function get_current_user_from_cookie in `src/api/v1/auth/auth.py`
+- Added template for orders panel
+
+### Changed
+- added env variables in `.env` file
+- bumping version in `src/__init__.py` file
+
+### Fixed
+- Enhanced error handling for 401 in `src/main.py`
+
 ## [1.5.1] 2026-05-23
 
 ### Added